+++ scripts/install_platform_with_component_manager.sh
@@ -1,21 +1,17 @@
 #!/usr/bin/env bash
 set -eo pipefail
 
-CURRENT_DIR=$(cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd)
-source "${CURRENT_DIR}/formats.sh"
-
-if [ -z "${SC_MACHINE_PATH}" ];
-then
-  source "${CURRENT_DIR}/set_vars.sh"
-fi
+SCRIPTS_DIR=$(cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd)
+SUBMODULE_SCRIPTS_DIR="${SCRIPTS_DIR}/submodule-scripts"
+source "${SCRIPTS_DIR}/formats.sh"
 
 stage "Install ostis-web-platform"
 
-"${PLATFORM_PATH}/scripts/install_sc_machine_submodule.sh"
-"${PLATFORM_PATH}/scripts/install_sc_machine_dependencies.sh" --dev
-"${PLATFORM_PATH}/scripts/install_sc_component_manager_submodule.sh"
-"${PLATFORM_PATH}/scripts/build_sc_machine.sh"
-"${PLATFORM_PATH}/scripts/build_sc_component_manager.sh"
-"${PLATFORM_PATH}/scripts/build_kb.sh"
+"${SUBMODULE_SCRIPTS_DIR}/install_sc_machine.sh"
+"${SCRIPTS_DIR}/install_sc_machine_dependencies.sh" --dev
+"${SUBMODULE_SCRIPTS_DIR}/install_sc_component_manager.sh"
+"${SCRIPTS_DIR}/build_sc_machine.sh"
+"${SCRIPTS_DIR}/build_sc_component_manager.sh"
+"${SCRIPTS_DIR}/build_kb.sh"
 
 stage "OSTIS-web-platform installed successfully"