+++ scripts/install_sc_machine_dependencies.sh
@@ -1,12 +1,12 @@
 #!/usr/bin/env bash
 set -eo pipefail
 
-CURRENT_DIR=$(cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd)
-source "${CURRENT_DIR}/formats.sh"
+SCRIPTS_DIR=$(cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd)
+source "${SCRIPTS_DIR}/formats.sh"
 
 if [ -z "${SC_MACHINE_PATH}" ];
 then
-  source "${CURRENT_DIR}/set_vars.sh"
+  source "${SCRIPTS_DIR}/set_vars.sh"
 fi
 
 stage "Install sc-machine dependencies"